Implement testing for Client side
Testing has been implemented to test the added classes that support the new user interface with the plant and bed lists.
Some getter methods were added to help with testing. These getters aren't used anywhere in the functional code.
The constructor for the plant class was modified. A better constructor (or builders) for the class should be made to initialize all fields.

Addresses issue #2, 3, 7