Implement changes from PlantInfoPages branch
The changes from the PlantInfoPages branch have been implemented and work in this branch.
The comments, like and dislike features all work and update to the database properly.
However, no link is shown in the browser when any of these features are used (I think this is just fine.)

There is quite a bit of refactoring and testing that needs to be done in the future. Issues will be added but ZenHub is non-functional right now. Will try again later tonight.